Hello Greg
Many thanks for your review!

> I have several suggestions for your consideration (purely editorial, as I think of them):
• in Abstract, it could be easier to use extended forms of acronyms like PREOF, OAM, DetNet, and PAREO

Can do, with the exception of DetNet which is more of a noun meaning our work rather than deterministic networking at large. PAREO I’d simply omit simpler that way.

• capitalize the title of Section 2.3.2 to "Recovery Graph"

Done

> capitalize "Forward" in the title of Section 2.3.3 (also capitalize in the first and second sentences of the first paragraph).

Done

> capitalize "Recovery Graph" in the title of Section 2.3.6. Also, if you decide to capitalize Complex, should "recovery graph" also be capitalized when used together?

Removed the “complex” thingy. A simple path is now a lane and that should be enough terminology.

> I think that the correct reference to the BFD specification is RFC 5880 (there seems like a typo referencing RFC 5580 in Section 2.6.7)

Oups

> "A PLR that Decides" reads like a PLR has a mind of its own ;). Would "A PLR that selects" or something similar be acceptable as a replacement?

Selects would have the same ias. We cold say follows the decision tree or performs the selection algorithm. But why discuss implementation? BTW the D is the one from OODA, that’s why it’s there.

> s/r CPF/rCPF/?

CPF is supposed to refer to the DetNet one. And r/aCPF to the RAW split of the CPF. I might have missed one instance, but that’s probably not a change all… “DetNet rCPF” should not have shown up; another change all artifact. I cleaned that.

> "sensitive/reactive" as a characterization of the DetNet rCPF seems like putting together sweet and overly sweet (in reverse order). Could both characteristics be replaced by "overreactive" or simply use "sensitive"?

What about

“
As a result, the DetNet CPF is not expected to be aware of and to
   react to very transient changes.
“
?
Many thanks again!

Pascal