Review by Rich Salz
Hello Rich:

Many thanks for your review, and sorry for the delay in responding.

Please see below:

I know nothing about routing, and wireless stuff, except to not trust the WiFi
in my local coffee shop.

The document gives a good overview about the need for and structure for
reliable wireless, and its layered approach.

I found Figure 1 confusing: the unlabeled lowercase "o" look like air bubbles
to me.  What are they and why are they relevant?

I would have loved to have a square character to represent a small device.
I used "o", changed to "#", not sure that helps. added a legend.

Similarly the "flakes of dandruff" in Figure 4 and the extra dots in Figure 8
and 10

Tried to make the clouds more cloudy :)

Having griped about the diagrams, I must say that they are a tour de force of
ASCII art.

:)

Focusing on the security considerations of Section 6 ...

"RAW uses all forms of diversity ..." do you mean transport?  Diversity as a
noun here seems wrong.

Now sure about how to use "diversity" here, being non-native speaker. For sure the RFC editor will help rephrase this .
In the meantime, to capture the intended meaning better I reworded as
"
   RAW leverages diversity (e.g., spatial and time diversity, coding
   diversity, and frequency diversity), possibly using heterogeneous
   wired and wireless networking technologies over different physical
   paths, to increase the reliability and availability in the face of
   unpredictable conditions.  While this is not done specifically to
   defeat an attacker, the amount of diversity used in RAW defeats
   possible attacks that would impact a particular technology or a
   specific path.
"

If Layer2 encryption prevents RAW component from promiscuous overhearing, what
is the next impact on the user?

Added
" Physical actions by a collocated attacker such as a radio
   interference may still lower the reliability of an end-to-end RAW
   transmission by blocking one segment or one possible path.  But if an
   alternate path with diverse frequency, location, and/or technology,
   is available, then RAW adapts by rerouting the impacted traffic over
   the preferred alternates, which defeats the attack after a limited
   period of lower reliability.  Then again, the security benefit is a
   side-effect of an action that is taken regardless of whether the
   source of the issue is voluntary (an attack) or not.
 "

but then this really belongs to the same group as the paragraph before this section, so I made that another section.

--
Pascal