Greg Mirsky's response to open area review
Hello Greg

GIM2>> Although some are confused by how "in-band" and "out-of-band" terms are used in the characterization of OAM methods, several WGs in the Routing Area, i.e., DetNet, NVO3, and BIER, accepted their interpretation in published documents. Furthermore, the statement that [I-D.ietf-opsawg-oam-characterization] is needed "to fully understand the semantics of the terms Active, Passive, Hybrid" seems inaccurate. RFC 7799 provides clear definitions and gives helpful examples for each class of OAM methods defined in RFC 7799. I believe that [I-D.ietf-opsawg-oam-characterization] adds no value regarding understanding these terms.

OPS Area insists on having that ref. I can see that your points make sense too. Let me try the below:

"
   RAW reuses terminology defined for Operations, Administration, and
   Maintenance (OAM) protocols in Section 1.1 of the "Framework of OAM
   for DetNet" [DetNet-OAM] and "Active and Passive Metrics and Methods
   (with Hybrid Types In-Between)" [RFC7799].  "Guidelines for
   Characterizing OAM" [I-D.ietf-opsawg-oam-characterization] provides
   additional semantics of the terms Active, Passive, Hybrid, and In-
   Packet OAM that are consistent with [RFC7799].  It also warns about
   potential inconsistencies in the way the terms "in-band" and "out-of-
   band" are used across the IETF; the DetNet reference for those terms
   is [DetNet-OAM].
 "

GIM2>> AFAICS, the DetNet WG accepted these terms and they are used in DetNet OAM documents without any controversy. I cannot find these terms being used in draft-ietf-raw-architecture. If that is the case, why do we need to discuss them in this document?

Because it is an architecture and docs that reference it will directly inherit the listed terminology, as opposed to indirectly via the references.
I agree they could refer to RFC 9551 and be done too, since in the case of OAM that particular doc exists. This is not the case for many other terms therein though.
I guess the main thing is not to introduce inconsistencies between this and its references. I hope the above text achieves that.