Éric Vyncke's No Objection on draft-ietf-raw-architecture-27: (with COMMENT)
Hello Eric

Many thanks for your review!

please see below:

----------------------------------------------------------------------
COMMENT:
----------------------------------------------------------------------

Please add an informative reference to `IEEE 802.1 Time-Sensitive networking`
and to `Time Slotted Channel Hopping (TSCH)`.

Actually the ref to RAW technologies is for both TSCH and OFDMA.

I changed to

"
                                                     With scheduled radios
   such as Time Slotted Channel Hopping (TSCH) and Orthogonal Frequency
   Division Multiple Access (OFDMA) (see [RAW-TECHNOS] for more on both
   of these and other technologies as well) being developed to provide
   determinism over wireless links at the lower layers, providing DetNet
   capabilities have become possible.
"

I am unable to parse `RAW and DetNet serve application flows that require a
special treatment along the paths that provide that treatment.`, please
consider re-phrasing it.

What about:
"
RAW and DetNet enable application flows that require a special treatment along paths that can provide that treatment.
"

Missing word(s) in `that are defined in [DetNet-OAM] but .` ?

fixed; the latest there is:
"

   RAW reuses terminology defined for Operations, Administration, and
   Maintenance (OAM) protocols in Section 1.1 of the "Framework of OAM
   for DetNet" [DetNet-OAM] and "Active and Passive Metrics and Methods
   (with Hybrid Types In-Between)" [RFC7799].  "Guidelines for
   Characterizing OAM" [I-D.ietf-opsawg-oam-characterization] provides
   additional semantics of the terms Active, Passive, Hybrid, and In-
   Packet OAM that are consistent with [RFC7799].  It also warns about
   potential inconsistencies in the way the terms "in-band" and "out-of-
   band" are used across the IETF; the DetNet reference for those terms
   is [DetNet-OAM].
"

Thanks for this useful terminology section.

Who is the `we` in `we refer to that complex scenario` ? The author ? the WG ?
the IETF community ? Please be accurate (or use the passive voice). Possiby in
other places as well.

changed to "this specification"

Should some reactions be added after `prompt detection of failures as they
occur`?

The expansion of that clause says "to direct maintenance". I believe that in the enumeration the short version is good enough.

Is there a need to introduce the FCAPS acronym ?

Not used by this spec but I thought useful knowledge for those who will build an implementation framework for this architecture

It is of course a difficult issue to fix w/o changing the flow, but referring
to figure that are much further down in the text is not easy for the reader.

Thanks for that. I moved that paragraph to 5.3 where it belongs.

Unsure whether the use of "wisdom" fits an IETF document in `generates
knowledge and wisdom`...

It's an obscure ref to the DIKW model. W -wisdom- is really AI at work. Changed both to a single 'intelligence'

s/which can be tuned to unsure /which can be tuned to *e*nsure / ?

oups, fixed

Like in all control loop, it is essential to aim for stability without
oscillation. The document seems to be light on this aspect.

 Very true: That's not really a problem with the architecture but with the rerouting algorithm. Still it should be mentioned in the problem statement. I added text as follows:

"
   RAW distinguishes the longer time-scale at which routes are computed
   from the shorter time-scale where forwarding decisions are made (see
   Section 6.1).  The RAW Network Plane operations happen at a time-
   scale that sits timewise between the routing and the forwarding time-
   scales.  Their goal is to select dynamically, within the resources
   delineated by a recovery graph, the protection path(s) that the
   upcoming packets of a DetNet flow shall follow.  As they influence
   the path for entire or portion of flows, the RAW Network Plane
   operations may affect the metrics used in their rerouting decision,
   which could potentially lead to oscillations; such effects must be
   avoided or dampened.
"

RAN in the figure appears before the expansion later in the text.

 moved the figure down slightly

s/. it/. It/ but I will stop noting these nits as the RFC Editor will make
another pass.

s/Operating at the Layer-3/Operating at the layer 3/ ?

removed the 'the'

To make a much nicer HTML rendering, suggest using the aasvg too to generate
SVG graphics. It is worth a try especially if the I-D uses the Kramdown file
format ;-)

I'll take a look. Using native xml.