Erik Kline's No Objection on draft-ietf-raw-architecture-27: (with COMMENT)
Hello Erik:

long time; many thanks for your review!

Please see below:

COMMENT:
----------------------------------------------------------------------

CC @ekline

* comment syntax:
  - https://github.com/mnot/ietf-comments/blob/main/format.md

* "Handling Ballot Positions":
  - https://ietf.org/about/groups/iesg/statements/handling-ballot-positions/

* "and hop count turns into a bad idea as the link budget drops with
   the distance"

  I think this means "drops with the physical distance".  If so, that
  might be a useful clarification as the mention of "hop count" can
  imply a different context for interpreting "distance" -- one that
  doesn't make any sense (i.e. link budget has nothing to do with the
  number of hops).

added "physical"

* "subset of the hops arof the packet rate observed"

  "arof"?  I couldn't parse this sentence in general.

some words were inserted there. Should be " just a subset of the hops are observed and controlled by RAW."

* This says that PER is the ratio of packets received in error
  (with errors?) to the total transmitted.  Should the denominator
  be total received?  Only the sender can know the total transmitted,
  as the receiver might not have received some packets at all (so
  no errors could be detected in a packet that never arrived).

True,  PDR and PER require to have an idea of the nb of attempts; to find that there's ARQ that can tell that a frame was missed, and there are the (turbo-line) FEC codes that analyse a frame (received, counted) and fail to recover the errors (so the frame is dropped). It's also possible through OAM to exchange figures and make computations.

all the best

Pascla

--
Pascal